requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pymysql==1.1.1
dbutils==3.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
